+++ compiler/src/main/java/org/neo4j/ogm/compiler/builders/node/DefaultNodeBuilder.java
@@ -11,7 +11,7 @@
  *  conditions of the subcomponent's license, as noted in the LICENSE file.
  */
 
-package org.neo4j.ogm.compiler.builders;
+package org.neo4j.ogm.compiler.builders.node;
 
 import java.util.Collection;
 import java.util.List;
