+++ compiler/src/main/java/org/neo4j/ogm/compiler/builders/statement/DeletedRelationshipStatementBuilder.java
@@ -11,7 +11,7 @@
  *  conditions of the subcomponent's license, as noted in the LICENSE file.
  */
 
-package org.neo4j.ogm.compiler.emitters;
+package org.neo4j.ogm.compiler.builders.statement;
 
 import java.util.*;
